+++ src/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
 
+## [v20190418]
+
+  - Pull minimum bet from contract, refs #129
+  - Fix permission issue on settings, refs #131
+
+
 ## [v20190217]
 
   - Migrate to upstream recipes
@@ -8,7 +14,7 @@
   - Split dedicated Etheroll library, refs #97
   - Remove legacy dependencies, refs #122
   - Move to python3 recipe, refs #123
-  - Handles write storage permission, refs #125
+  - Handle write storage permission, refs #125
   - Non root Docker user, refs #127